Update patch set 2
Patch Set 2:

Phil, Your change here was correct. That is why I accepted it :-)

As I mentioned in my followup comment on https://gerrit.iotivity.org/gerrit/#/c/16209/
the specific build error (only for RIOT-OS native builds) I was trying to address needs a different fix which I will soon push. But the other change to clock.c in 16209 was necessary since RIOT-OS changed their xtimer API. As a result of 16209, the samples when built from port/riot/Makefile now build successfully.

As for the comments on https://github.com/RIOT-OS/RIOT/issues/6241, it appears to be because the package itself is behind and needs maintenance. 
I have just posted a comment over there, and they can choose to assign it to the package author @mattiantonini, or I will take a look at it myself when I get a chance.
However, that issue wasn't related to change 16209 as you assumed. 

Hope this resolves the confusion!

Patch-set: 2