Update patch set 1
Patch Set 1: Code-Review-1

I've made a comment about the sleep()'s usage in the other commit. If that comment doesn't not make sense, please rebase it with that commit, there is no reason to have this fix for something that is not in the upstream yet.

Patch-set: 1
Reviewer: Gerrit User 211 <211@c6509b8a-a471-4ec1-8e5f-9aed2654acb5>
Label: Code-Review=-1