Update patch set 2
Patch Set 2:

(7 comments)

Besides the comments, this dynamic allocation is something that we really need (mainly when linux is target). One thing that this patch does not cover is the buffer size used by cbor (g_encoder). Easily we are facing problems with it, we need to check when an overflow happens in cbor and then try to alloc more memory. Of course, this can be address further.

Patch-set: 2
Label: Verified=0