Update patch set 2
Patch Set 2:

(6 comments)

> One thing that this patch does not cover is the buffer size used by cbor 
> (g_encoder). 

Well, this is what the MAX_APP_DATA_SIZE parameter is for. It specifies the maximum size of the OCF app/CoAP payload (which is entirely CBOR encoded). If we want to send/receive large payloads then we must raise that limit. Ofcourse, doing this on a static build would allocate large buffers.

Patch-set: 2
Label: Verified=0