Update patch set 5
Patch Set 5:

The Linux and Windows ports are for development only and are not meant for production. I imagine developers may want to use IPv4 when testing because it's possible their VMs and/or their default Linux installations don't support IPv6 or their corporate networks block it.

On production, IPv6 is mandatory. That means a certified device that enables IPv4 must also have IPv6. Also, IPv6 is sufficient: since all devices must have it, communication can happen over IPv6.

So explain to me: why should a constrained device carry IPv4 *and* IPv6, when all communication on production will be on IPv6?

Patch-set: 5