Update patch set 2
Patch Set 2:

Hi, thanks for feedback

I agrea this could be relocated, but I prefered to align the most to iotivity scripts, so It would be more convenient to have those files in root (or changed in iotivity but maybe it's too late now).

note the script can also be used to developers for do more advanced checks (or reproduce CI's behavior locally).

Once verifiers are working we can eventually create a task for refactoring all, but I don't want to drift too much on what has been done and (working with iotivity) as I am not familiar enough with jbb.

Patch-set: 2