Update patch set 2
Patch Set 2:

Hi,

To be clear, I would like to see these scripts go under tests/ and do not want it 
at the root level. It is simple enough for anyone to run all the test related scripts from tests/.

The rationale is to retain the current logical organization of the source and build
methodology (going into port/os/ to do a build), and having any top-level build
scripts serves to confuse.

The pattern we use here does not have to exactly match IoTivity.

Patch-set: 2
Label: Verified=0