Update patch set 1
Patch Set 1:

Andreas,

Updating the code for the 64-bit build is almost all about casing from a 64-bit value size_t to an int.

For most of the code it makes since to just replace the int with size_t. There are a lot of functions that are using the negative value of the int to return an error status. In those instances I have been casting. See the change in 'api/oc_rep.c' for an example.

Another option would be to cast this to ssize_t so it still represents the size of an allocated block of memory but can indicate an error with the -1.  I am actually leaning toward using the ssize_t value.

There are also sections in the code that I felt should be set to size_t but when I changed them to size_t they would fail the 32-bit build because size_t was not declared in scope. I would need to include stdlib.h or stddef.h the problem with including these headers I think they were purposely left out of these headers to keep the code size small as possible. Example of this is the _oc_mem_alloc is currently defined to take `unsigned int` for the memory size. When I change it to `size_t` it will fail the 32-bit build.

One of the big errors that I am getting in the 64-bit build is

    warning C4244: 'function': conversion from 'SOCKET' to 'int', possible loss of data

In the 64-bit build SOCKET is a 64-bit value. Based on my (very limited) research SOCKET currently will always return a value that is 0 to 2^24 which will fit in an int. This is just what currently happens there no contract keeping it in this range.

See this stackoverflow question and read all the comments for the accepted answer. https://stackoverflow.com/questions/1953639/is-it-safe-to-cast-socket-to-int-under-win64

I really don't know if its good practice to cast those to an int or should I be looking for another solution?

Patch-set: 1
Label: Verified=0