Update patch set 1
Patch Set 1:

I have updated the build instructions that are part of the Readme document that is downloaded with Iotivity-lite. This was done in response to the work done by the dmtools_tg to try and improve usability for developers.

Patch-set: 1
Reviewer: Gerrit User 4407 <4407@c6509b8a-a471-4ec1-8e5f-9aed2654acb5>
Label: Verified=0