Update patch set 1
Patch Set 1:

As far as I can tell this should be working but it is inconsistent. every so often the test fails which is typically a result of improper memory management and garbage collection.

I wanted to share it but I don't expect you guys to understand the change.

Patch-set: 1
Reviewer: Gerrit User 4407 <4407@c6509b8a-a471-4ec1-8e5f-9aed2654acb5>
Label: Verified=0