Update patch set 4
Patch Set 4:

I have done a lot of testing on this for windows. This makes it so the out put can be used in C or C++.

I have tested in only on windows using C

I would like to change the actual compilation to C in different  commit that can be tested in isolation to this change.

Patch-set: 4